# Conflicts:
#	packages/react-native/React/Fabric/Mounting/ComponentViews/TextInput/RCTTextInputComponentView.mm
#	packages/rn-tester-e2e/tests/helpers/utils.js
#	packages/rn-tester-e2e/tests/screens/components.screen.js

+++ packages/react-native/ReactCommon/react/nativemodule/core/platform/ios/ReactCommon/RCTInteropTurboModule.h
@@ -23,7 +23,7 @@ class JSI_EXPORT ObjCInteropTurboModule : public ObjCTurboModule {
   struct MethodDescriptor {
     std::string methodName;
     SEL selector;
-    int jsArgCount;
+    size_t jsArgCount;
     TurboModuleMethodValueKind jsReturnKind;
   };
